moves hooks and assets
