+++ app/layout.tsx
@@ -1,6 +1,6 @@
 import "../src/styles/globals.css";
 import { Metadata } from "next";
-import { Layout } from "@/components/Layout";
+import { Layout } from "@components/Layout";
 import { Poppins, Montserrat } from "next/font/google";
 
 export const metadata: Metadata = {
