+++ app/loading.tsx
@@ -1,4 +1,4 @@
-import { LoadingSpinner } from "@/components/LoadingSpinner";
+import { LoadingSpinner } from "@components/LoadingSpinner";
 
 export default function Loading() {
   return <LoadingSpinner />;
