+++ src/components/LinedPaper/LinedPaper.module.css
@@ -1,4 +1,3 @@
-
 /**
 HEADER
 ---------------------
@@ -18,7 +17,7 @@ EVERY OTHER LINE
 **/
 .paper {
   --left-gap: 2rem;
-  --header-gap: calc(var(--line-height)* 2);
+  --header-gap: calc(var(--line-height) * 2);
   --colored-line-size: 2px;
 
   position: relative;
@@ -38,7 +37,7 @@ EVERY OTHER LINE
   opacity: 1;
 }
 .paper::before {
-  content: '';
+  content: "";
   position: absolute;
   top: 0;
   bottom: 0;
@@ -47,7 +46,7 @@ EVERY OTHER LINE
   border-right: var(--colored-line-size) solid var(--red-20);
 }
 .paper::after {
-  content: '';
+  content: "";
   position: absolute;
   top: var(--header-gap);
   bottom: 0;
@@ -57,17 +56,17 @@ EVERY OTHER LINE
 
   /* repeated pattern to draw lined paper lines */
   background: linear-gradient(
-                var(--blue-20) var(--colored-line-size),
-                transparent var(--colored-line-size),
-                transparent calc(var(--line-height) - var(--colored-line-size))
-              );
+    var(--blue-20) var(--colored-line-size),
+    transparent var(--colored-line-size),
+    transparent calc(var(--line-height) - var(--colored-line-size))
+  );
 
   background-size: auto var(--line-height);
 }
 
 /* if there is no title, add padding to preserve effect */
 .paper.noTitle {
-  padding-top: var(--header-gap); 
+  padding-top: var(--header-gap);
 }
 
 .paper > header:not(.multiLine) > .title {
@@ -80,6 +79,9 @@ EVERY OTHER LINE
   margin-bottom: var(--line-height);
 }
 
-.paper h3, .paper h4, .paper h5, .paper h6 {
+.paper h3,
+.paper h4,
+.paper h5,
+.paper h6 {
   font-weight: bold;
 }
