+++ src/hooks/useStyle/index.ts
@@ -1,0 +1,1 @@
+export { useStyle } from "./useStyle";
